requirements.txt
pygame-ce==2.5.3
PyGLM
PyOpenGL
esper==2.2
